data-generator/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
numpy==1.24.3
faker==18.10.1
